rename assets/assets.dart -> assets/index.dart
